+++ rmi/src/main/kotlin/org/pastalab/fray/rmi/ScheduleObserver.kt
@@ -1,0 +1,18 @@
+package org.pastalab.fray.rmi
+
+import java.rmi.Remote
+import java.rmi.RemoteException
+
+interface ScheduleObserver<T> : Remote {
+  @Throws(RemoteException::class) fun onExecutionStart()
+
+  @Throws(RemoteException::class) fun onNewSchedule(enabledSchedules: List<T>, scheduled: T)
+
+  @Throws(RemoteException::class) fun onExecutionDone()
+
+  @Throws(RemoteException::class) fun saveToReportFolder(path: String)
+
+  companion object {
+    const val NAME = "ScheduleObserver"
+  }
+}